user authetication aded with login, logout and registartion

+++ myapp/views.py
@@ -1,5 +1,10 @@
-from unicodedata import name
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate,login,logout
 from .models import Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
@@ -13,6 +18,53 @@ from django.db.models import Q
 # ]
 
 # Create your views here.
+
+def loginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request,'User does not Exists!!')
+
+        user = authenticate(request,username=username,password=password)
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'Username or Password does not Exists!!')
+            
+
+    context={'page':page}
+    return render(request,'login_register.html',context)
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    context={'form':form}
+    return render(request,'login_register.html',context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     
@@ -32,6 +84,7 @@ def room(request,pk):
     context = {"room":room}
     return render(request, "room.html", context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST' :
@@ -43,10 +96,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
     
+    if request.user != room.host:
+        return HttpResponse("You are not allowed")
+
     if request.method == 'POST' :
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -58,9 +115,13 @@ def updateRoom(request, pk):
     
     return render(request,'room_form.html',context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed")
+
 
     if request.method == 'POST' :
         room.delete()
